exporter/requirements.txt
prometheus-client==0.8.0
PyYAML==5.3.1
